simplified_llm_api_script/requirements.txt
anthropic
httpx
openai
python-frontmatter
pydantic
python-dotenv
python-docx
beautifulsoup4
tiktoken
